clean up parse logic and adds app reducer state
