+++ frontend/scripts/react-components/tool-links/tool-links.serializers.js
@@ -1,0 +1,18 @@
+export const detailedView = {
+  parse(param) {
+    return JSON.parse(param);
+  }
+};
+
+export const selectedResizeBy = {
+  stringify(prop, DONT_SERIALIZE) {
+    if (!prop) {
+      return DONT_SERIALIZE;
+    }
+
+    return prop.name;
+  },
+  parse(param) {
+    return { name: param };
+  }
+};
