+++ frontend/scripts/react-components/tool/tool.js
@@ -3,10 +3,11 @@ import ToolComponent from 'react-components/tool/tool.component';
 import { connect } from 'react-redux';
 import { resizeSankeyTool } from 'react-components/tool/tool.thunks';
 import { getToolLinksUrlProps } from 'react-components/tool-links/tool-links.selectors';
+import { getAppUrlProps } from 'reducers/app.selectors';
 
 const getUrlProps = createSelector(
-  [getToolLinksUrlProps],
-  toolLinks => ({ ...toolLinks })
+  [getToolLinksUrlProps, getAppUrlProps],
+  (toolLinks, app) => ({ ...toolLinks, ...app })
 );
 
 const mapStateToProps = state => ({
